requirements.txt
Flask
Flask-Cors
flask-sqlalchemy
psycopg2
pysqlite3
sqlalchemy-pysqlite3
numpy
numpydoc
scikit-image
scikit-learn
scipy
gunicorn
